Move `Cache` generation to separate module
